Add files via upload
This is a script to be used with an Azure Application to pull device names from devices in Intune and then tag the device in Defender for Endpoint to create an easy connection between the iOS device. This is because iOS name in Intune and what is in Defender is most likely going to be different.